+++ power.betfair/Classes/cancelInstruction.ps1
@@ -1,0 +1,4 @@
+class cancelInstruction {
+    [string] $betId
+    [double] $sizeReduction
+}