+++ power.betfair/Public/Get-BetfairMarketType.ps1
@@ -29,7 +29,7 @@ Function Get-BetfairMarketType {
     Restrict markets by the venue associated with the market. Currently, only Horse & Greyhound racing markets have venues.
 
     .PARAMETER bspOnly
-    Restrict to bsp markets only, if True or non-bsp markets if False. If not specified then returns both BSP and non-BSP markets
+    Restrict to bsp markets only, if True or non-bsp markets if False. If not specified then returns both BSP and non-BSP markets.
 
     .PARAMETER turnInPlayEnabled
     Restrict to markets that will turn in play if True or will not turn in play if false. If not specified, returns both.
@@ -38,7 +38,7 @@ Function Get-BetfairMarketType {
     Restrict to markets that are currently in play if True or are not currently in play if false. If not specified, returns both.
 
     .PARAMETER marketBettingTypes
-    Restrict to markets that match the betting type of the market (i.e. Odds, Asian Handicap Singles, Asian Handicap Doubles or Line)
+    Restrict to markets that match the betting type of the market (i.e. Odds, Asian Handicap Singles, Asian Handicap Doubles or Line).
 
     .PARAMETER marketCountries
     Restrict to markets that are in the specified country or countries.  Please note: the default value is 'GB' when the correct country code cannot be determined.
@@ -46,10 +46,10 @@ Function Get-BetfairMarketType {
     .PARAMETER marketTypeCodes
     Restrict to markets that match the type of the market (i.e., MATCH_ODDS, HALF_TIME_SCORE).
     You should use this instead of relying on the market name as the market type codes are the same in all locales.
-    Please note: All market types are available via the listMarketTypes operations
+    Please note: All market types are available via the listMarketTypes operations.
 
     .PARAMETER marketStartTime
-    Restrict to markets with a market start time before or after the specified date
+    Restrict to markets with a market start time before or after the specified date.
 
     .PARAMETER withOrders
     Restrict to markets where I have one or more orders in these states.
@@ -64,7 +64,7 @@ Function Get-BetfairMarketType {
     Get-BetfairMarketType -textQuery "MATCH_ODDS"
 
     .NOTES
-    General notes
+    General notes.
     #>
 
     [CmdletBinding()][OutputType('System.Management.Automation.PSObject')]